release on 2015-11-06_08-35-46

+++ src/lagrangian/cfdemParticle/etc/functions.sh
@@ -80,13 +80,21 @@ compileLib()
                 echo "Please make sure to have the compressible libraries first in the library-list.txt!"
                 cd $CFDEM_SRC_DIR/lagrangian/cfdemParticle
                 echo "changing to $PWD"
-                rmdepall 2>&1 | tee -a $logpath/$logfileName
+                if [[ $WM_PROJECT_VERSION == "dev" ]]; then
+                    wrmdep 2>&1 | tee -a $logpath/$logfileName
+                else
+                    rmdepall 2>&1 | tee -a $logpath/$logfileName
+                fi
                 cd $casePath
                 echo "changing to $PWD"
             else
                 echo "Compiling a incompressible library."
         fi
-        rmdepall 2>&1 | tee -a $logpath/$logfileName
+        if [[ $WM_PROJECT_VERSION == "dev" ]]; then
+            wrmdep 2>&1 | tee -a $logpath/$logfileName
+        else
+            rmdepall 2>&1 | tee -a $logpath/$logfileName
+        fi
         wclean 2>&1 | tee -a $logpath/$logfileName
     #fi
     wmake libso 2>&1 | tee -a $logpath/$logfileName
@@ -126,11 +134,18 @@ compileSolver()
     pwd 2>&1 | tee -a $logpath/$logfileName
     echo 2>&1 | tee -a $logpath/$logfileName
 
-    #- wclean and wmake
-    #if [ $doClean != "noClean" ]; then
-        rmdepall 2>&1 | tee -a $logpath/$logfileName
-        wclean 2>&1 | tee -a $logpath/$logfileName
-    #fi
+    # check if there is an Allwmake
+    if [ -f "Allwmake" ]; then
+        echo "doing Allwclean and Allwmake for the solver"
+        bash Allwclean 2>&1 | tee -a $logpath/$logfileName
+        bash Allwmake 2>&1 | tee -a $logpath/$logfileName
+    else
+        #- wclean and wmake
+        #if [ $doClean != "noClean" ]; then
+            rmdepall 2>&1 | tee -a $logpath/$logfileName
+            wclean 2>&1 | tee -a $logpath/$logfileName
+        #fi
+    fi
     
     # compile parallel?
     if [[ $parallel == "true" ]]; then
@@ -178,7 +193,7 @@ compileLIGGGHTS()
     else
         rm $CFDEM_LIGGGHTS_SRC_DIR/"lmp_"$CFDEM_LIGGGHTS_MAKEFILE_NAME
         rm $CFDEM_LIGGGHTS_SRC_DIR/"lib"$CFDEM_LIGGGHTS_LIB_NAME".a"
-        make clean-all 2>&1 | tee -a $logpath/$logfileName
+        make clean-$CFDEM_LIGGGHTS_MAKEFILE_NAME 2>&1 | tee -a $logpath/$logfileName
         echo "cleaning LIGGGHTS"
     fi
     if [[ $WM_NCOMPPROCS == "" ]]; then
@@ -544,7 +559,21 @@ parCFDrun()
     rm $logpath/$logfileName
 
     #- change path
-    cd $casePath/CFD
+    cd $casePath
+
+    #- remove old data
+    rm -rf processor*
+
+    #- decompose case
+    decomposePar
+
+    #- make proc dirs visible
+    count=0
+    for i in `seq $nrProcs`
+    do
+        let count=$i-1
+        (cd $casePath/processor$count && touch file.foam)
+    done
 
     #- header
     echo 2>&1 | tee -a /$logpath/$logfileName
@@ -557,7 +586,7 @@ parCFDrun()
 
     #- run applictaion
     if [ $machineFileName == "none" ]; then
-        mpirun -np $nrProcs $solverName -parallel 2>&1 | tee -a $logpath/$logfileName
+        mpirun -np $nrProcs $debugMode $solverName -parallel 2>&1 | tee -a $logpath/$logfileName
     else
         mpirun -machinefile $machineFileName -np $nrProcs $debugMode $solverName -parallel 2>&1 | tee -a $logpath/$logfileName
     fi
@@ -694,7 +723,7 @@ collectLogCFDEMcoupling_sol()
     LASTWORD=$(basename $LASTSTRING)
 
     # log if compilation was success  
-    if [[ $LASTWORD == $SOLVERNAME ]]; then
+    if [[ $LASTWORD == $SOLVERNAME || $LASTWORD == "date." ]]; then
         echo "$SOLVERNAME" >> $logpath/log_compile_results_success
     else
         echo "$SOLVERNAME" >> $logpath/log_compile_results_fail
